refactor(DPGO-19): divide handlers into topics
user and password topic, generic util funcs and validator
