+++ internal/server/http/handlers/message.go
@@ -1,0 +1,22 @@
+package handlers
+
+import (
+	"encoding/json"
+	"net/http"
+)
+
+// Message represent error message.
+type Message struct {
+	Msg string
+}
+
+func WriteErrorResponse(code int, msg string, w http.ResponseWriter) {
+	w.WriteHeader(code)
+	json.NewEncoder(w).Encode(Message{msg})
+}
+
+func Render(w http.ResponseWriter, data interface{}) {
+	enc := json.NewEncoder(w)
+	enc.SetIndent("", "\t")
+	enc.Encode(data)
+}
